+++ src/modules/claims/claims.service.ts
@@ -7,7 +7,10 @@ import {
   PreconditionType,
   RoleCredentialSubject,
 } from '@energyweb/credential-governance';
-import { VerifiableCredential } from '@ew-did-registry/credentials-interface';
+import {
+  VerifiableCredential,
+  VerifiablePresentation,
+} from '@ew-did-registry/credentials-interface';
 import { ClaimRevocation } from '@energyweb/onchain-claims';
 import { Methods } from '@ew-did-registry/did';
 import { Algorithms } from '@ew-did-registry/jwt';
@@ -67,9 +70,10 @@ import {
 import {
   CredentialResolver,
   EthersProviderIssuerResolver,
+  EthersProviderRevokerResolver,
   IpfsCredentialResolver,
-  VCIssuerVerification,
-  ClaimIssuerVerification,
+  IssuerVerification,
+  RevocationVerification,
   RoleEIP191JWT,
 } from '@energyweb/vc-verification';
 import { DidRegistry } from '../did-registry/did-registry.service';
@@ -104,7 +108,7 @@ export class ClaimsService {
   private _claimManager: string;
   private _claimManagerInterface = ClaimManager__factory.createInterface();
   private _claimRevocation: ClaimRevocation;
-  private _vcIssuerVerifier: VCIssuerVerification;
+  private _issuerVerification: IssuerVerification;
   private _issuerResolver: EthersProviderIssuerResolver;
   private _credentialResolver: CredentialResolver;
   private _statusVerifier: StatusListEntryVerification;
@@ -117,7 +121,7 @@ export class ClaimsService {
     private _verifiableCredentialService: VerifiableCredentialsServiceBase
   ) {
     this._signerService.onInit(this.init.bind(this));
-    this._setClaimIssuerVerifier();
+    this._setIssuerVerifier();
     this._setStatusVerifier();
   }
 
@@ -471,6 +475,16 @@ export class ClaimsService {
 
     if (registrationTypes.includes(RegistrationTypes.OffChain)) {
       await this.verifyIssuer(claimData.claimType);
+      const vp = await this.issueVerifiablePresentation({
+        subject: sub,
+        namespace: role,
+        version: version.toString(),
+        issuerFields,
+        credentialStatus: credentialStatusOverride,
+        expirationTimestamp: claimExpirationTimestamp,
+      });
+      const vpCredentialStatus = vp?.verifiableCredential[0]?.credentialStatus;
+      const credentialStatus = credentialStatusOverride || vpCredentialStatus;
       const publicClaim: IPublicClaim = {
         did: sub,
         signer: this._signerService.did,
@@ -478,24 +492,14 @@ export class ClaimsService {
           ...strippedClaimData,
           ...(issuerFields && { issuerFields }),
         },
-        //credentialStatus: credentialStatus
+        ...(credentialStatus && { credentialStatus }),
       };
-      const [issuedToken, vp] = await Promise.all([
-        this._didRegistry.issuePublicClaim({
-          publicClaim,
-          expirationTimestamp: claimExpirationTimestamp,
-        }),
-        this.issueVerifiablePresentation({
-          subject: sub,
-          namespace: role,
-          version: version.toString(),
-          issuerFields,
-          credentialStatus: credentialStatusOverride,
-          expirationTimestamp: claimExpirationTimestamp,
-        }),
-      ]);
+      const issuedToken = await this._didRegistry.issuePublicClaim({
+        publicClaim,
+        expirationTimestamp: claimExpirationTimestamp,
+      });
       message.issuedToken = issuedToken;
-      message.vp = vp;
+      message.vp = JSON.stringify(vp);
     }
 
     await this._cacheClient.issueClaim(this._signerService.did, message);
@@ -664,28 +668,28 @@ export class ClaimsService {
     };
 
     if (registrationTypes.includes(RegistrationTypes.OffChain)) {
+      const vp = await this.issueVerifiablePresentation({
+        subject,
+        namespace: claim.claimType,
+        version: claim.claimTypeVersion.toString(),
+        issuerFields: claim.issuerFields,
+        credentialStatus: credentialStatusOverride,
+        expirationTimestamp: claimExpirationTimestamp,
+      });
+      const vpCredentialStatus = vp?.verifiableCredential[0]?.credentialStatus;
+      const credentialStatus = credentialStatusOverride || vpCredentialStatus;
       const publicClaim: IPublicClaim = {
         did: subject,
         signer: this._signerService.did,
         claimData: claim,
+        ...(credentialStatus && { credentialStatus }),
       };
-      const [issuedToken, vp] = await Promise.all([
-        this._didRegistry.issuePublicClaim({
-          publicClaim,
-          expirationTimestamp: claimExpirationTimestamp,
-        }),
-        this.issueVerifiablePresentation({
-          subject,
-          namespace: claim.claimType,
-          version: claim.claimTypeVersion.toString(),
-          issuerFields: claim.issuerFields,
-          credentialStatus: credentialStatusOverride,
-          expirationTimestamp: claimExpirationTimestamp,
-        }),
-      ]);
-
+      const issuedToken = await this._didRegistry.issuePublicClaim({
+        publicClaim,
+        expirationTimestamp: claimExpirationTimestamp,
+      });
       message.issuedToken = issuedToken;
-      message.vp = vp;
+      message.vp = JSON.stringify(vp);
     }
 
     if (registrationTypes.includes(RegistrationTypes.OnChain)) {
@@ -1226,7 +1230,7 @@ export class ClaimsService {
    * @param {String} role Registration types of the claim
    */
   private async verifyIssuer(role: string): Promise<void> {
-    await this._vcIssuerVerifier.verifyIssuer(this._signerService.did, role);
+    await this._issuerVerification.verifyIssuer(this._signerService.did, role);
   }
 
   /**
@@ -1278,7 +1282,7 @@ export class ClaimsService {
    */
   private async issueVerifiablePresentation(
     options: IssueVerifiablePresentationOptions
-  ): Promise<string> {
+  ): Promise<VerifiablePresentation> {
     const vc = await this._verifiableCredentialService.createRoleVC({
       id: options.subject,
       namespace: options.namespace,
@@ -1289,11 +1293,9 @@ export class ClaimsService {
         ? new Date(options.expirationTimestamp)
         : undefined,
     });
-    const vp =
-      await this._verifiableCredentialService.createVerifiablePresentation([
-        vc,
-      ]);
-    return JSON.stringify(vp);
+    return await this._verifiableCredentialService.createVerifiablePresentation(
+      [vc]
+    );
   }
 
   /**
@@ -1447,7 +1449,7 @@ export class ClaimsService {
     const role = vc.credentialSubject.role.namespace;
     let issuerVerified = true;
     try {
-      await this._vcIssuerVerifier.verifyIssuer(issuerDID, role);
+      await this._issuerVerification.verifyIssuer(issuerDID, role);
     } catch (e) {
       issuerVerified = false;
       errors.push((e as Error).message);
@@ -1481,13 +1483,7 @@ export class ClaimsService {
     const { payload, eip191Jwt } = roleEIP191JWT;
     const errors: string[] = [];
     const issuerDID = this._signerService.did;
-    const claimIssuerVerifier = new ClaimIssuerVerification(
-      this._signerService.provider,
-      this._didRegistry.registrySettings,
-      this._credentialResolver,
-      this._issuerResolver
-    );
-    const issuerVerified = await claimIssuerVerifier.verifyIssuer(
+    const issuerVerified = await this._issuerVerification.verifyIssuer(
       issuerDID,
       payload?.claimData?.claimType
     );
@@ -1541,7 +1537,7 @@ export class ClaimsService {
    * Set the Verifier for Claim Issuance.
    *
    */
-  private _setClaimIssuerVerifier() {
+  private _setIssuerVerifier() {
     this._credentialResolver = new IpfsCredentialResolver(
       this._signerService.provider,
       this._didRegistry.registrySettings,
@@ -1549,14 +1545,27 @@ export class ClaimsService {
     );
     const domainReader = this._domainsService.domainReader;
     this._issuerResolver = new EthersProviderIssuerResolver(domainReader);
-    this._vcIssuerVerifier = new VCIssuerVerification(
+    const revokerResolver = new EthersProviderRevokerResolver(domainReader);
+    const verifyProof = (vc: string, proofOptions: string) =>
+      this._verifiableCredentialService.verify(
+        JSON.parse(vc),
+        JSON.parse(proofOptions)
+      );
+    const revocationVerification = new RevocationVerification(
+      revokerResolver,
       this._issuerResolver,
       this._credentialResolver,
-      (vc: string, proofOptions: string) =>
-        this._verifiableCredentialService.verify(
-          JSON.parse(vc),
-          JSON.parse(proofOptions)
-        )
+      this._signerService.provider,
+      this._didRegistry.registrySettings,
+      verifyProof
+    );
+    this._issuerVerification = new IssuerVerification(
+      this._issuerResolver,
+      this._credentialResolver,
+      this._signerService.provider,
+      this._didRegistry.registrySettings,
+      revocationVerification,
+      verifyProof
     );
   }